python3/pyproject.toml
[tool.poetry]
name = "python3"
version = "0.1.0"
description = ""
authors = ["Шарафиев Ришат <rsharafiev@rbc.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
confluent-kafka = "^1.5.0"
python-decouple = "^3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
